refactor(step-2): extract MCP response builders into src/tools/responses.ts
Moves createToolResponse and createErrorResponse verbatim out of utils.ts
into src/tools/responses.ts — the first file in the new tools/ layer.
Signatures and behavior unchanged (same Pick<CallToolResult, ...> return
types, same annotations audience). Updates main.ts and utils.ts (which
still holds handleNoteTextUpdate temporarily) to import from the new
location.
