+++ src/utils.ts
@@ -4,6 +4,7 @@ import { CORE_DATA_EPOCH_OFFSET } from './config.js';
 import { getNoteContent } from './notes.js';
 import { buildBearUrl, executeBearXCallbackApi } from './bear-urls.js';
 import { logAndThrow, logger } from './logging.js';
+import { createErrorResponse, createToolResponse } from './tools/responses.js';
 
 /**
  * Decodes and normalizes Bear tag names.
@@ -109,38 +110,6 @@ export function parseDateString(dateString: string): Date {
       return parsed;
     }
   }
-}
-
-/**
- * Creates a standardized MCP tool response with consistent formatting.
- * Centralizes response structure to follow DRY principles.
- *
- * @param text - The response text content
- * @returns Formatted CallToolResult for MCP tools
- */
-export function createToolResponse(text: string): Pick<CallToolResult, 'content'> {
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text,
-        annotations: { audience: ['user', 'assistant'] as const },
-      },
-    ],
-  };
-}
-
-/**
- * Creates a standardized MCP error response with isError flag.
- * Signals to the LLM that the tool failed and self-correction may be needed.
- * Per the MCP spec, tool errors use isError: true inside the result object
- * so the LLM can see the failure and retry or adjust its approach.
- *
- * @param text - The error description with recovery guidance
- * @returns Formatted CallToolResult with isError: true
- */
-export function createErrorResponse(text: string): Pick<CallToolResult, 'content' | 'isError'> {
-  return { ...createToolResponse(text), isError: true };
 }
 
 /**
